Break out the main `App` and `NonceApp` data types into a separate module.
This makes the `What4.Expr.Builder` module somewhat easier to navigate,
and may help reduce build times slightly.  To accomplish this change,
some API changes were necessary.  Most notable is the change to
`definedFn`, which now takes a datatype defining the unfolding policy
instead of a user-provided callback function.
