+++ what4/what4.cabal
@@ -132,6 +132,9 @@ library
 
     Test.Verification
 
+  other-modules:
+    What4.Expr.App
+
   ghc-options: -Wall -Werror=incomplete-patterns -Werror=missing-methods -Werror=overlapping-patterns
   ghc-prof-options: -O2 -fprof-auto-top
   if impl(ghc >= 8.6)
